Add permbot_agent cmd, an in-cluster applier
This changes around the way Permbot works, to allow it to work in a more
"gitops" fashion - where it watches a set of configmaps (applied by e.g
Weaveworks Flux) to the cluster.

Rather than forcing users to explicitly write out
{,Cluster}Role{,Binding}s, the system can pop the TOML into a configmap
and permbot will notice and turn it into the applicable K8S objects.